backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.11
django-debug-toolbar==4.3.0
django-filter==24.2
django-stubs==5.0.0
django-stubs-ext==5.0.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.0
djoser==2.2.2
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.0.0
gunicorn==21.2.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
mypy==1.10.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tomli==2.0.1
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.20240406
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1